update: move GamePage component to the @views/pages namespace
