+++ src/views/pages/GamePage/GamePage.module.scss
