+++ src/views/pages/GamePage/GamePage.tsx
