+++ src/views/pages/GamePage/IGamePage.ts
