+++ src/views/pages/GamePage/index.ts
