+++ src/views/routes/index.ts
@@ -1,3 +1,2 @@
 export * from './Root/Root'
-export * from './Game/GamePage'
 export * from './Search/Search'
